fix(discriminator): widen to fp32 before the STFT
The trainer calls the critic from inside the autocast block, so with the
default mixed_precision: bf16 the reconstruction reaches STFTDiscriminator._stft
as bf16 — and torch.stft has no reduced-precision kernel. Every adversarial
step would have raised the moment adv_start_step was passed on a real GPU run.

Widen before analysing and let autocast re-cast for the convolutions, matching
what the loss path already does. Covered by a test that feeds the critic bf16
directly.

+++ synthgen/model/discriminator.py
@@ -103,7 +103,11 @@ class STFTDiscriminator(nn.Module):
 
     def _stft(self, x: torch.Tensor) -> torch.Tensor:
         """Return the complex STFT as a real 2-channel image."""
-        # x: (batch * channels, samples)
+        # x: (batch * channels, samples).
+        # torch.stft has no reduced-precision kernel, and the trainer calls the
+        # critic from inside an autocast block, so the input arrives as bf16.
+        # Analyse in fp32 and let autocast re-cast for the convolutions.
+        x = x.float()
         spec = torch.stft(
             x,
             n_fft=self.n_fft,
